Fixing train

+++ leaffliction/augmentations.py
@@ -135,6 +135,9 @@ class AugmentationEngine:
                     suffix
                 )
                 pm.ensure_dir(augm_path.parent)
+                if augm_path.exists():
+                    print(f"⏭️  Skipped existing file: {augm_path}")
+                    continue
                 cv2.imwrite(str(augm_path), transformed_image)
                 augmented_items.append((augm_path, item[1]))
